tweak: separate oracle test files
