--- test/012_oracles_other.ts
@@ -1,173 +1,0 @@
-import { constants, id } from '@yield-protocol/utils-v2'
-const { WAD } = constants
-import { CHI, RATE } from '../src/constants'
-
-import { sendStatic } from './shared/helpers'
-
-import { Contract } from '@ethersproject/contracts'
-import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/dist/src/signer-with-address'
-
-import OracleArtifact from '../artifacts/contracts/mocks/oracles/OracleMock.sol/OracleMock.json'
-import CompoundMultiOracleArtifact from '../artifacts/contracts/oracles/compound/CompoundMultiOracle.sol/CompoundMultiOracle.json'
-import CTokenMultiOracleArtifact from '../artifacts/contracts/oracles/compound/CTokenMultiOracle.sol/CTokenMultiOracle.json'
-import CTokenChiMockArtifact from '../artifacts/contracts/mocks/oracles/compound/CTokenChiMock.sol/CTokenChiMock.json'
-import CTokenRateMockArtifact from '../artifacts/contracts/mocks/oracles/compound/CTokenRateMock.sol/CTokenRateMock.json'
-import CDaiMockArtifact from '../artifacts/contracts/mocks/oracles/compound/CDaiMock.sol/CDaiMock.json'
-import CUSDCMockArtifact from '../artifacts/contracts/mocks/oracles/compound/CUSDCMock.sol/CUSDCMock.json'
-import UniswapV3FactoryMockArtifact from '../artifacts/contracts/mocks/oracles/uniswap/UniswapV3FactoryMock.sol/UniswapV3FactoryMock.json'
-import UniswapV3OracleArtifact from '../artifacts/contracts/oracles/uniswap/UniswapV3Oracle.sol/UniswapV3Oracle.json'
-
-import { IOracle } from '../typechain/IOracle'
-import { CompoundMultiOracle } from '../typechain/CompoundMultiOracle'
-import { CTokenMultiOracle } from '../typechain/CTokenMultiOracle'
-import { CTokenChiMock } from '../typechain/CTokenChiMock'
-import { CTokenRateMock } from '../typechain/CTokenRateMock'
-import { CDaiMock } from '../typechain/CDaiMock'
-import { CUSDCMock } from '../typechain/CUSDCMock'
-import { UniswapV3FactoryMock } from '../typechain/UniswapV3FactoryMock'
-import { UniswapV3PoolMock } from '../typechain/UniswapV3PoolMock'
-import { UniswapV3Oracle } from '../typechain/UniswapV3Oracle'
-
-import { ethers, waffle } from 'hardhat'
-import { expect } from 'chai'
-const { deployContract } = waffle
-
-function bytes6ToBytes32(x: string): string {
-  return x + '00'.repeat(26)
-}
-
-describe('Oracles - Compound/cToken/Uniswap', function () {
-  this.timeout(0)
-
-  let ownerAcc: SignerWithAddress
-  let owner: string
-  let oracle: IOracle
-  let compoundMultiOracle: CompoundMultiOracle
-  let cTokenMultiOracle: CTokenMultiOracle
-  let cTokenChi: CTokenChiMock
-  let cTokenRate: CTokenRateMock
-  let cDai: CDaiMock
-  let cUSDC: CUSDCMock
-  let uniswapV3Factory: UniswapV3FactoryMock
-  let uniswapV3Pool: UniswapV3PoolMock
-  let uniswapV3PoolAddress: string
-  let uniswapV3Oracle: UniswapV3Oracle
-
-  const baseId = ethers.utils.hexlify(ethers.utils.randomBytes(6))
-  const usdcId = ethers.utils.hexlify(ethers.utils.randomBytes(6))
-  const cDaiId = ethers.utils.hexlify(ethers.utils.randomBytes(6))
-  const cUSDCId = ethers.utils.hexlify(ethers.utils.randomBytes(6))
-  const usdQuoteId = ethers.utils.hexlify(ethers.utils.randomBytes(6))
-  const ethQuoteId = ethers.utils.hexlify(ethers.utils.randomBytes(6))
-  const mockBytes6 = ethers.utils.hexlify(ethers.utils.randomBytes(6))
-  const mockBytes32 = ethers.utils.hexlify(ethers.utils.randomBytes(32))
-
-  before(async () => {
-    const signers = await ethers.getSigners()
-    ownerAcc = signers[0]
-    owner = await ownerAcc.getAddress()
-  })
-
-  beforeEach(async () => {
-    oracle = (await deployContract(ownerAcc, OracleArtifact, [])) as IOracle
-
-    cTokenChi = (await deployContract(ownerAcc, CTokenChiMockArtifact, [])) as CTokenChiMock
-    cTokenRate = (await deployContract(ownerAcc, CTokenRateMockArtifact, [])) as CTokenRateMock
-    cDai = (await deployContract(ownerAcc, CDaiMockArtifact, [])) as CDaiMock
-    cUSDC = (await deployContract(ownerAcc, CUSDCMockArtifact, [])) as CUSDCMock
-
-    compoundMultiOracle = (await deployContract(ownerAcc, CompoundMultiOracleArtifact, [])) as CompoundMultiOracle
-    await compoundMultiOracle.grantRole(id('setSource(bytes6,bytes6,address)'), owner)
-    await compoundMultiOracle.setSource(baseId, CHI, cTokenChi.address)
-    await compoundMultiOracle.setSource(baseId, RATE, cTokenRate.address)
-
-    cTokenMultiOracle = (await deployContract(ownerAcc, CTokenMultiOracleArtifact, [])) as CTokenMultiOracle
-    await cTokenMultiOracle.grantRole(id('setSource(bytes6,bytes6,address)'), owner)
-    await cTokenMultiOracle.setSource(cDaiId, baseId, cDai.address)
-    await cTokenMultiOracle.setSource(cUSDCId, usdcId, cUSDC.address)
-
-    uniswapV3Factory = (await deployContract(ownerAcc, UniswapV3FactoryMockArtifact, [])) as UniswapV3FactoryMock
-    const token0: string = ethers.utils.HDNode.fromSeed('0x0123456789abcdef0123456789abcdef').address
-    const token1: string = ethers.utils.HDNode.fromSeed('0xfedcba9876543210fedcba9876543210').address
-    uniswapV3PoolAddress = await sendStatic(uniswapV3Factory as Contract, 'createPool', ownerAcc, [token0, token1, 0])
-    uniswapV3Pool = (await ethers.getContractAt('UniswapV3PoolMock', uniswapV3PoolAddress)) as UniswapV3PoolMock
-    uniswapV3Oracle = (await deployContract(ownerAcc, UniswapV3OracleArtifact, [])) as UniswapV3Oracle
-    await uniswapV3Oracle.grantRole(id('setSource(bytes6,bytes6,address)'), owner)
-    await uniswapV3Oracle.setSource(baseId, ethQuoteId, uniswapV3PoolAddress)
-  })
-
-  it('sets and retrieves the value at spot price', async () => {
-    await oracle.set(WAD.mul(2))
-    expect((await oracle.callStatic.get(mockBytes32, mockBytes32, WAD))[0]).to.equal(WAD.mul(2))
-  })
-
-  it('revert on unknown sources', async () => {
-    await expect(
-      compoundMultiOracle.callStatic.get(bytes6ToBytes32(mockBytes6), bytes6ToBytes32(mockBytes6), WAD)
-    ).to.be.revertedWith('Source not found')
-    await expect(
-      cTokenMultiOracle.callStatic.get(bytes6ToBytes32(mockBytes6), bytes6ToBytes32(mockBytes6), WAD)
-    ).to.be.revertedWith('Source not found')
-    await expect(
-      uniswapV3Oracle.callStatic.get(bytes6ToBytes32(mockBytes6), bytes6ToBytes32(mockBytes6), WAD)
-    ).to.be.revertedWith('Source not found')
-  })
-
-  it('sets and retrieves the chi and rate values at spot price from a compound multioracle', async () => {
-    await cTokenChi.set(WAD.mul(2))
-    await cTokenRate.set(WAD.mul(3))
-    expect((await compoundMultiOracle.callStatic.get(bytes6ToBytes32(baseId), bytes6ToBytes32(CHI), WAD))[0]).to.equal(
-      WAD.mul(2)
-    )
-    expect((await compoundMultiOracle.callStatic.get(bytes6ToBytes32(baseId), bytes6ToBytes32(RATE), WAD))[0]).to.equal(
-      WAD.mul(3)
-    )
-
-    expect((await compoundMultiOracle.peek(bytes6ToBytes32(baseId), bytes6ToBytes32(CHI), WAD))[0]).to.equal(WAD.mul(2))
-    expect((await compoundMultiOracle.peek(bytes6ToBytes32(baseId), bytes6ToBytes32(RATE), WAD))[0]).to.equal(
-      WAD.mul(3)
-    )
-  })
-
-  it('sets and retrieves the cToken spot price from a cToken multioracle', async () => {
-    await cDai.set(WAD.mul(2).mul(10 ** 10)) // cDai has 18 + 10 decimals
-    await cUSDC.set(WAD.mul(2).div(100)) // USDC has 6 + 10 decimals
-
-    expect((await cTokenMultiOracle.callStatic.get(bytes6ToBytes32(cDaiId), bytes6ToBytes32(baseId), WAD))[0]).to.equal(
-      WAD.mul(2)
-    )
-    expect(
-      (await cTokenMultiOracle.callStatic.get(bytes6ToBytes32(cUSDCId), bytes6ToBytes32(usdcId), WAD))[0]
-    ).to.equal(WAD.mul(2))
-    expect((await cTokenMultiOracle.callStatic.get(bytes6ToBytes32(baseId), bytes6ToBytes32(cDaiId), WAD))[0]).to.equal(
-      WAD.div(2)
-    )
-
-    expect((await cTokenMultiOracle.peek(bytes6ToBytes32(cDaiId), bytes6ToBytes32(baseId), WAD))[0]).to.equal(
-      WAD.mul(2)
-    )
-    expect((await cTokenMultiOracle.peek(bytes6ToBytes32(cUSDCId), bytes6ToBytes32(usdcId), WAD))[0]).to.equal(
-      WAD.mul(2)
-    )
-    expect((await cTokenMultiOracle.peek(bytes6ToBytes32(baseId), bytes6ToBytes32(cDaiId), WAD))[0]).to.equal(
-      WAD.div(2)
-    )
-  })
-
-  it('retrieves the value at spot price from a uniswap v3 oracle', async () => {
-    await uniswapV3Pool.set(WAD.mul(2))
-    expect(
-      (await uniswapV3Oracle.callStatic.get(bytes6ToBytes32(baseId), bytes6ToBytes32(ethQuoteId), WAD))[0]
-    ).to.equal(WAD.mul(2))
-    expect(
-      (await uniswapV3Oracle.callStatic.get(bytes6ToBytes32(ethQuoteId), bytes6ToBytes32(baseId), WAD))[0]
-    ).to.equal(WAD.div(2))
-
-    expect((await uniswapV3Oracle.peek(bytes6ToBytes32(baseId), bytes6ToBytes32(ethQuoteId), WAD))[0]).to.equal(
-      WAD.mul(2)
-    )
-    expect((await uniswapV3Oracle.peek(bytes6ToBytes32(ethQuoteId), bytes6ToBytes32(baseId), WAD))[0]).to.equal(
-      WAD.div(2)
-    )
-  })
-})
